requirements.txt
trueskill
customtkinter
pillow